requirements.txt
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
biopython @ file:///D:/bld/biopython_1768646503356/work
cloudpickle==3.1.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy @ file:///D:/bld/contourpy_1744743084386/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_cycler_1764466758/work
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1769745019/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1763918002538/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fair-esm==2.0.0
filelock==3.20.3
fonttools @ file:///D:/bld/fonttools_1765632560086/work
fsspec==2026.1.0
gensim==3.4.0
imbalanced-learn @ file:///home/conda/feedstock_root/build_artifacts/imbalanced-learn_1766355322435/work
ipykernel @ file:///D:/bld/bld/rattler-build_ipykernel_1761567966/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1748711254/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1765794027227/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_client_1767954655/work
jupyter_core @ file:///D:/bld/bld/rattler-build_jupyter_core_1760643889/work
kiwisolver @ file:///D:/bld/bld/rattler-build_kiwisolver_1762488749/work
llvmlite==0.46.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mpmath==1.3.0
munkres==1.1.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
numba==0.63.1
numpy @ file:///D:/bld/numpy_1747544591878/work/dist/numpy-2.2.6-cp310-cp310-win_amd64.whl#sha256=af66576ce916c795ffa81be9620c8138a3fda6e1713b2bf7775a630e47eca3a8
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
pandas @ file:///D:/bld/pandas_1764614482805/work
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
patsy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_patsy_1760998269/work
peptides==0.5.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow @ file:///D:/bld/bld/rattler-build_pillow_1758208682/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1764950726/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
propy3==1.1.1
propythia==3.0.2
psutil @ file:///D:/bld/bld/rattler-build_psutil_1769678167/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==23.0.0
pyfasta==0.5.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyparsing_1769003998/work
PySide6==6.9.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487147/work
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1757387021/work
scikit-learn @ file:///D:/bld/scikit-learn_1757432814257/work/dist/scikit_learn-1.7.2-cp310-cp310-win_amd64.whl#sha256=80780e7bac7c11947bc8f76f74cce4a729277e8fc2c67f76db5038e95de040c9
scipy @ file:///C:/bld/scipy-split_1739790833536/work/dist/scipy-1.15.2-cp310-cp310-win_amd64.whl#sha256=4748853a0c723cda1fa2bdead847ba7c2443bde9cec3436e3e7243cfa54b55ff
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1733730015268/work
shap==0.49.1
shiboken6==6.9.2
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sklearn-compat @ file:///home/conda/feedstock_root/build_artifacts/sklearn-compat_1766329668603/work
slicer==0.0.8
smart_open==7.5.0
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
statsmodels @ file:///D:/bld/statsmodels_1764983000605/work
sympy==1.14.0
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
torch==2.11.0.dev20260128+cu128
torchaudio==2.11.0.dev20260128+cu128
torchvision==0.25.0.dev20260128+cu128
tornado @ file:///D:/bld/tornado_1765836387919/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1765719872007/work
unicodedata2 @ file:///D:/bld/unicodedata2_1763054693050/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1769858610910/work
wrapt==2.0.1
xgboost @ file:///home/conda/feedstock_root/build_artifacts/xgboost-split_1768314074650/work/python-package
